requirements.txt
Flask==2.0.1
Flask-Caching==1.10.1
redis==3.5.3
waitress==2.0.0
pika==1.2.0
threaded==4.1.0
python-dotenv==0.19.0